Write a function that returns the sum of all its parameters.
    Prototype: int sum_them_all(const unsigned int n, ...);
    If n == 0, return 0

+++ variadic_functions/0-sum_them_all.c
@@ -3,13 +3,13 @@
 /**
  * sum_them_all - function that returns the sum of all its parameters
  * @n: number of args
- * Return: sum
+ * Return: sum (or 0 if n == 0)
 */
 int sum_them_all(const unsigned int n, ...)
 {
 	va_list args;
 	unsigned int i;
-	int sum = 0;
+	double sum = 0;
 
 	if (n == 0)
 		return (0);
